LayoutTests/js/exception-properties-expected.txt

Test for correct properties on Error objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.keys(error).sort() is []
PASS Object.keys(nativeError).sort() is []
PASS Object.getOwnPropertyNames(error).sort() is ["column", "line", "message", "sourceURL", "stack"]
PASS Object.getOwnPropertyNames(nativeError).sort() is ["column", "line", "message", "sourceURL", "stack"]
PASS Object.getPrototypeOf(nativeError).name is "RangeError"
PASS Object.getPrototypeOf(nativeError).message is ""
PASS successfullyParsed is true

TEST COMPLETE
